backend/requirements.txt
fastapi==0.122.0
openai-agents==0.6.1
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
